requirements.txt
Flask>=3.0,<4.0
Flask-Caching>=2.0,<3.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Login>=0.6,<1.0
Flask-WTF>=1.2,<2.0
python-dotenv>=1.0,<2.0
PyMySQL>=1.1,<2.0
Werkzeug>=3.0,<4.0
gunicorn>=21.0,<22.0
cryptography
openpyxl>=3.1,<4.0
